\name{regulatoR-package}
\alias{regulatoR-package}
\alias{regulatoR}
\docType{package}
\title{
  regulatoR
}
\description{
  a package for regulatoRy inference
}
\details{
\tabular{ll}{
Package: \tab regulatoR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-06-14\cr
License: \tab GPL\cr
}
}
\author{
Tim Triche, Jr. <tim.triche@gmail.com>
}
\keyword{ package }
\examples{
  # library(regulatoR)
}
